Use API to get members
Since when listing a channel, the list of users may contain all of them,
some of them or none of them, completely at random, said list is absolutely
useless.

So I am instead just using the num_members for the /list command, and using a
separate API to get the ids of the members of a channel.

Fun fact: also num_members is buggy and can be 0 for no reason whatsoever.

At slack they really realy know how to make a public facing API.