Add --autojoin option: force join remote chans.
This is closer to the original Slack gateway: a user is automatically
joined to all its server-side channels (public, private) and multi-user
IMs.

Note: we suffer from a very low performance for this along with
userlist: fetching each user profile is slow; and should be optimized in
a init time mass loading.