Unable to handle file_deleted, so ignoring it
It does not carry enough information to do anything about it at all.

The indicated file id is not valid at that point, so it is not possible to know
which user is deleting what, so I could only show "file deleted" but can't say
the user. Kinda pointless. The link won't work anyway.