startup: support environment variables for PORT and TOKEN.
Due to the one to one mapping of service instances to listen ports /
tokens, being able to pull the most commonly adjusted config arguments
straight out of the environment will be really handy, especially for
those of us who use a lot of process supervision.

This necessitated moving token file parsing out of slack.py and into
irc.py in addition to modifications to the argument handling.

Something to keep in mind: as written this commit changes precedence so
that if PORT or TOKEN are set their corresponding optargs are ignored.
To switch that order around, add a second test to see if args.token or
args.port are set to their defaults and only apply the environment value
if the arg values are still default.