prefix username mentions with '@'
When username mentions are done in a message you will see '@username'
instead of 'username'. Also sending a message with a username mention now
requires you to type '@username'.

This makes it more clear when someone is mentioned. It also fixes the
issue where for example you have a user called 'are'. Every time you type
'are' this user was mentioned.