Improve the history feature
The last message seen is sent again with the history, since it has the last
seen timestamp.

So this commit ignores that message.

Also to avoid the history to start again from the same point in time at
the next reconnection, if no new messages were sent, this considers
the timestamps in the history messages as the latest timestamp so
that they won't be requested again to slack.