Do not insert mentions in URLs
Prior to this mentions were inserted looking for them with a
regex like <word-separator>username<word-separator>

This worked but in the case of http.//username/ it would match.

The new regex searches for an initial and optional sequence of "://"
(to match any url schema) followed by any number of non space characters
and then with the old regex at the end.

So URLs still match but their match is different so they can be skipped.

This further changes the match, instead of creating one regex per username
only one giant one is created, putting in OR all the patterns for all the
usernames.

Because the regex is only ran once the substitution is performed
right to left, in order to preserve the position of the matches.

Closes: #81