Remove rchat support
It is a bit sad but I haven't used it in a while and it becomes unlikely that
it can work after switching to asyncio without testing it.

Nobody stepped up to do any work so eh. I'll keep the source
file around just in case I need to add it back.