Only tag users from the same channel
Fixes: https://github.com/ltworf/localslackirc/issues/217

Instead of adding all of the users present on the slack instance to the
regexp to perform substitutions, create the regexp on the fly with only the
users that are present in the destination channel.

This also changes the behaviour in private messages, where now tagging
usernames will not be done at all.