Show edits by word rather than by letter
When showing edits, it will proceed until a word boundary is met (or the string
ends) rather than only showing the chars that got changed.

So if a user fixes a typo in a word, the entire word that got fixed is shown,
so one can understand what is being edited with more clarity.

Fixes: #205