irc: improve URL formatting
The current URL formatting makes messages with more than one link barely
readable. This changes the formatting by using just the message label in
the body of the message with a reference number (¹, ² etc). The URLs are
appended to the end of message, prefixed with a newline, two spaces and the
corresponding reference number. The two spaces are added to help
separate that message from the next ones.

Before:

<bot> [GitLab] Antonio Terceiro (terceiro) https://gitlab.com/Linaro/tuxmake/-/issues/131#note_544882634 (commented on issue #131) in https://gitlab.com/Linaro/tuxmake (linaro / tuxmake): *socat server fails sometimes*

After:

<bot> [GitLab] Antonio Terceiro (terceiro) commented on issue #131¹ in linaro / tuxmake²: *socat server fails sometimes*
<bot>   ¹ https://gitlab.com/Linaro/tuxmake/-/issues/131#note_544882634
<bot>   ² https://gitlab.com/Linaro/tuxmake