Refactor handling of attached files
Basically the file shared event contains information about which channel is receiving the file but not
if the file is referenced into a thread or in the main channel.

This changes a bunch of things.

No longer handle the FileShared event at all, but instead read the files field in messages and
get the links from there.

So the announce is no longer needed to create a fake message and file_shared events are ignored.

Doing this, automatically the shared files end up in their right place.

Also the call to get file information is no longer needed and was removed.

Closes: #342