Only get the full list of channels when needed
In environments with too many channels, getting them gets rate limited.

With this change, history and autojoin list only fetch the list of joined channels.

The full list is still fetched when /list runs or when parsing messages which refer to
channels which we haven't joined.

In those cases, errors are ignored and the list of whatever was obtained is returned, so that
hopefully the needed channel was already fetched.

Closes: #440